+++ src/types/theme/BaseZIndex.ts
@@ -5,12 +5,8 @@ export type BaseZIndex = {
     behind: number
     base: number
     dropdown: number
-    sticky: number
-    fixed: number
     overlay: number
     modal: number
-    popover: number
     tooltip: number
-    toast: number
 }