+++ src/utils/theme/defaultTheme.ts
@@ -79,13 +79,9 @@ export const defaultZIndex: Theme['zIndex'] = {
     behind: -1,
     base: 0,
     dropdown: 1000,
-    sticky: 1100,
-    fixed: 1200,
     overlay: 1300,
     modal: 1400,
-    popover: 1500,
     tooltip: 1600,
-    toast: 1700,
 }
 
 /**